+++ tests/testthat/test-probdist.R
@@ -1,0 +1,122 @@
+context("probdist class")
+
+genparm <- function(type = "int") {
+  # ad-hoc function to randomly generate probability distribution parameters
+  parm <- switch(type,
+    "int"  = rpois(1L, lambda = 10L),
+    "prob" = runif(1L),
+    "norm" = rnorm(1L, mean = 0L, sd = 10L),
+    "pos"  = rexp(1L, rate = .2),
+    "neg"  = -genparm("pos"),
+    stop("invalid type")
+  )
+  return(parm)
+}
+
+test_that("Beta parameteres are properly converted", {
+  fam <- "beta"
+  s1 <- genparm()
+  s2 <- genparm()
+
+  prbdst <- probdist(shape1 = s1, shape2 = s2, family = fam)
+  expect_equal(prbdst$family, fam)
+  expect_equal(prbdst$parms, c(shape1 = s1, shape2 = s2))
+  expect_equal(prbdst$nat_parms, c(eta1 = s1, eta2 = s2))
+
+  prbdst_nat <- probdist(eta1 = s1, eta2 = s2, family = fam)
+  expect_equal(prbdst_nat$family, fam)
+  expect_equal(prbdst_nat$parms, c(shape1 = s1, shape2 = s2))
+  expect_equal(prbdst_nat$nat_parms, c(eta1 = s1, eta2 = s2))
+})
+
+test_that("Binomial parameteres are properly converted", {
+  fam <- "binomial"
+  sz <- genparm()
+  pb <- genparm("prob")
+  et <- genparm("norm")
+
+  prbdst <- probdist(size = sz, prob = pb, family = fam)
+  expect_equal(prbdst$family, fam)
+  expect_equal(prbdst$parms, c(size = sz, prob = pb))
+  expect_equal(c(prbdst$nat_parms), c(eta = log(pb / (1 - pb))))
+
+  prbdst_nat <- probdist(eta = et, family = fam)
+  expect_equal(prbdst_nat$family, fam)
+  expect_equal(prbdst_nat$parms, c(prob = exp(et) / (1 + exp(et))))
+  expect_equal(prbdst_nat$nat_parms, c(eta = et))
+})
+
+test_that("Chisq parameteres are properly converted", {
+  fam <- "chisq"
+  nu <- genparm()
+  et <- genparm()
+
+  prbdst <- probdist(df = nu, family = fam)
+  expect_equal(prbdst$family, fam)
+  expect_equal(prbdst$parms, c(df = nu))
+  expect_equal(prbdst$nat_parms, c(eta = nu / 2 - 1))
+
+  prbdst_nat <- probdist(eta = et, family = fam)
+  expect_equal(prbdst_nat$family, fam)
+  expect_equal(prbdst_nat$parms, c(df = 2 * (et + 1)))
+  expect_equal(prbdst_nat$nat_parms, c(eta = et))
+})
+
+test_that("Contbern parameteres are properly converted", {
+  fam <- "contbern"
+  lb <- genparm("prob")
+  et <- genparm("norm")
+
+  prbdst <- probdist(lambda = lb, family = fam)
+  expect_equal(prbdst$family, fam)
+  expect_equal(prbdst$parms, c(lambda = lb))
+  expect_equal(prbdst$nat_parms, c(eta = log(lb / (1 - lb))))
+
+  prbdst_nat <- probdist(eta = et, family = fam)
+  expect_equal(prbdst_nat$family, fam)
+  expect_equal(prbdst_nat$parms, c(lambda = exp(et) / (1 + exp(et))))
+  expect_equal(prbdst_nat$nat_parms, c(eta = et))
+})
+
+test_that("Exponential parameteres are properly converted", {
+  fam <- "exp"
+  rt <- genparm("pos")
+  et <- genparm("neg")
+
+  prbdst <- probdist(rate = rt, family = fam)
+  expect_equal(prbdst$family, fam)
+  expect_equal(prbdst$parms, c(rate = rt))
+  expect_equal(prbdst$nat_parms, c(eta = -rt))
+
+  prbdst_nat <- probdist(eta = et, family = fam)
+  expect_equal(prbdst_nat$family, fam)
+  expect_equal(prbdst_nat$parms, c(rate = -et))
+  expect_equal(prbdst_nat$nat_parms, c(eta = et))
+})
+
+test_that("Gamma parameteres are properly converted", {
+  fam <- "gamma"
+  alpha <- genparm("pos")
+  beta <- genparm("pos")
+  k <- genparm("pos")
+  theta <- genparm("pos")
+  et1 <- genparm("norm")
+  et2 <- genparm("norm")
+
+  prbdst_1 <- probdist(shape = k, scale = theta, family = fam)
+  expect_equal(prbdst_1$family, fam)
+  expect_equal(prbdst_1$parms, c(shape = k, scale = theta))
+  expect_equal(prbdst_1$nat_parms, c(eta1 = k - 1, eta2 = -1 / theta))
+
+  prbdst_2 <- probdist(shape = alpha, rate = beta, family = fam)
+  expect_equal(prbdst_2$family, fam)
+  expect_equal(prbdst_2$parms, c(shape = alpha, rate = beta))
+  expect_equal(prbdst_2$nat_parms, c(eta1 = alpha - 1, eta2 = -beta))
+
+  prbdst_nat <- probdist(eta1 = et1, eta2 = et2, family = fam)
+  expect_equal(prbdst_nat$family, fam)
+  expect_equal(prbdst_nat$parms, c(shape = et1 + 1, rate = -et2))
+  expect_equal(prbdst_nat$nat_parms, c(eta1 = et1, eta2 = et2))
+})
+
+test_that("Errors are properly caught", {})